+++ cxx_mkspecs/gxx_common.py
@@ -143,5 +143,5 @@ def mkspec_set_gxx_cxxflags(conf):
     if conf.has_tool_option('cxx_nodebug'):
         conf.env['DEFINES'] += ['NDEBUG']
 
-    # Use the C++14 language features
-    conf.env['CXXFLAGS'] += ['-std=c++14']
+    # Use the C++17 language features
+    conf.env['CXXFLAGS'] += ['-std=c++17']