+++ cxx_mkspecs/msvc_common.py
@@ -123,6 +123,9 @@ def mkspec_set_msvc_flags(conf):
         ['/O2', '/W2', '/wd4503', '/wd4312',
          '/EHs', '/D_WIN32_WINNT=0x0501', '/bigobj']
 
+    # Use the C++17 language features
+    conf.env['CXXFLAGS'] += ['-std:c++17']
+
     # Do not generate .manifest files (the /MANIFEST flag is added by waf)
     conf.env['LINKFLAGS'].remove('/MANIFEST')
     conf.env['LINKFLAGS'] += ['/MANIFEST:NO']